+++ src/main/java/edu/seu/housepricepredict/domain/pojo/year/CommunityYearPrice.java
@@ -1,31 +1,31 @@
-package edu.seu.housepricepredict.pojo.year;
+package edu.seu.housepricepredict.domain.pojo.year;
 
 /**
  * @author guodonwu@163.com
- * @date 14:22 2019/2/27
- * 街道历史年份房价
+ * @date 14:21 2019/2/27
+ * 小区历史年份房价
  */
 
-public class StreetYearPrice {
-    private int sId;
+public class CommunityYearPrice {
+    private int coId;
     private int year;
     private int price;
 
     @Override
     public String toString() {
-        return "StreetYearPrice{" +
-                "sId=" + sId +
+        return "CommunityYearPrice{" +
+                "coId=" + coId +
                 ", year=" + year +
                 ", price=" + price +
                 '}';
     }
 
-    public int getsId() {
-        return sId;
+    public int getCoId() {
+        return coId;
     }
 
-    public void setsId(int sId) {
-        this.sId = sId;
+    public void setCoId(int coId) {
+        this.coId = coId;
     }
 
     public int getYear() {
