+++ src/main/java/edu/seu/housepricepredict/mapper/CommunityMapper.java
@@ -7,4 +7,6 @@ package edu.seu.housepricepredict.mapper;
  */
 
 public interface CommunityMapper {
+
+
 }
